DCS-76 Extracting out statement client
