+++ server/services/statementService.js
@@ -3,18 +3,18 @@ const { StatementStatus } = require('../config/types')
 const statementConfig = require('../config/statement')
 const { validate } = require('../utils/fieldValidation')
 
-module.exports = function createReportService({ incidentClient }) {
-  const getStatementsForUser = async (userId, status) => {
-    const data = await incidentClient.getStatementsForUser(userId, status)
+module.exports = function createStatementService({ statementsClient }) {
+  const getStatements = async (userId, status) => {
+    const data = await statementsClient.getStatements(userId, status)
     return data.rows
   }
 
   const getStatement = async (userId, reportId, status) => {
-    const statement = await incidentClient.getStatement(userId, reportId, status)
+    const statement = await statementsClient.getStatement(userId, reportId, status)
     if (!statement) {
       throw new Error(`Report: '${reportId}' does not exist`)
     }
-    const additionalComments = await incidentClient.getAdditionalComments(statement.id)
+    const additionalComments = await statementsClient.getAdditionalComments(statement.id)
     return { additionalComments, ...statement }
   }
 
@@ -26,22 +26,22 @@ module.exports = function createReportService({ incidentClient }) {
 
   const save = (userId, reportId, statement) => {
     logger.info(`Saving statement for user: ${userId} and report: ${reportId}`)
-    return incidentClient.saveStatement(userId, reportId, statement)
+    return statementsClient.saveStatement(userId, reportId, statement)
   }
 
   const submitStatement = (userId, reportId) => {
     logger.info(`Submitting statement for user: ${userId} and report: ${reportId}`)
-    return incidentClient.submitStatement(userId, reportId)
+    return statementsClient.submitStatement(userId, reportId)
   }
 
   const saveAdditionalComment = (statementId, additionalComment) => {
     logger.info(`Saving additional comment for statement with id: ${statementId}`)
-    return incidentClient.saveAdditionalComment(statementId, additionalComment)
+    return statementsClient.saveAdditionalComment(statementId, additionalComment)
   }
 
   return {
     getStatement,
-    getStatementsForUser,
+    getStatements,
     save,
     submitStatement,
     validateSavedStatement,
