+++ server/services/statementService.test.js
@@ -1,9 +1,8 @@
 const serviceCreator = require('./statementService')
 const { StatementStatus } = require('../config/types')
 
-const incidentClient = {
-  getCurrentDraft: jest.fn(),
-  getStatementsForUser: jest.fn(),
+const statementsClient = {
+  getStatements: jest.fn(),
   getStatement: jest.fn(),
   saveStatement: jest.fn(),
   submitStatement: jest.fn(),
@@ -14,10 +13,9 @@ const incidentClient = {
 let service
 
 beforeEach(() => {
-  service = serviceCreator({ incidentClient })
-  incidentClient.getCurrentDraft.mockReturnValue({ a: 'b' })
-  incidentClient.getStatementsForUser.mockReturnValue({ rows: [{ id: 1 }, { id: 2 }] })
-  incidentClient.saveAdditionalComment(50, 'Some new comment')
+  service = serviceCreator({ statementsClient })
+  statementsClient.getStatements.mockReturnValue({ rows: [{ id: 1 }, { id: 2 }] })
+  statementsClient.saveAdditionalComment(50, 'Some new comment')
 })
 
 afterEach(() => {
@@ -27,18 +25,18 @@ afterEach(() => {
 describe('saveAdditionalComment', () => {
   test('save a new comment', async () => {
     await service.saveAdditionalComment(50, 'Some new comment')
-    expect(incidentClient.saveAdditionalComment).toBeCalledWith(50, 'Some new comment')
+    expect(statementsClient.saveAdditionalComment).toBeCalledWith(50, 'Some new comment')
   })
 })
 
-describe('getStatementsForUser', () => {
+describe('getStatements', () => {
   test('retrieve  details', async () => {
-    const output = await service.getStatementsForUser('user1', StatementStatus.PENDING)
+    const output = await service.getStatements('user1', StatementStatus.PENDING)
 
     expect(output).toEqual([{ id: 1 }, { id: 2 }])
 
-    expect(incidentClient.getStatementsForUser).toBeCalledTimes(1)
-    expect(incidentClient.getStatementsForUser).toBeCalledWith('user1', StatementStatus.PENDING)
+    expect(statementsClient.getStatements).toBeCalledTimes(1)
+    expect(statementsClient.getStatements).toBeCalledWith('user1', StatementStatus.PENDING)
   })
 })
 
@@ -46,8 +44,8 @@ describe('getStatement', () => {
   test('retrieve details when user is involved', async () => {
     const statement = { id: 1, user_id: 'BOB' }
     const comment = 'Some additional text'
-    incidentClient.getStatement.mockReturnValue(statement)
-    incidentClient.getAdditionalComments.mockReturnValue([
+    statementsClient.getStatement.mockReturnValue(statement)
+    statementsClient.getAdditionalComments.mockReturnValue([
       {
         id: 1,
         statement_id: 1,
@@ -70,15 +68,15 @@ describe('getStatement', () => {
       ],
     })
 
-    expect(incidentClient.getStatement).toBeCalledWith('BOB', 1, StatementStatus.PENDING)
+    expect(statementsClient.getStatement).toBeCalledWith('BOB', 1, StatementStatus.PENDING)
   })
 
   test('retrieve details when statement is not present', async () => {
-    incidentClient.getStatement.mockReturnValue(undefined)
+    statementsClient.getStatement.mockReturnValue(undefined)
 
     await expect(service.getStatement('BOB', 1, 'PENDING')).rejects.toThrow(new Error("Report: '1' does not exist"))
 
-    expect(incidentClient.getStatement).toBeCalledWith('BOB', 1, 'PENDING')
+    expect(statementsClient.getStatement).toBeCalledWith('BOB', 1, 'PENDING')
   })
 })
 
@@ -87,20 +85,20 @@ test('should call save', async () => {
 
   await service.save('user1', 'incident-1', statement)
 
-  expect(incidentClient.saveStatement).toBeCalledTimes(1)
-  expect(incidentClient.saveStatement).toBeCalledWith('user1', 'incident-1', statement)
+  expect(statementsClient.saveStatement).toBeCalledTimes(1)
+  expect(statementsClient.saveStatement).toBeCalledWith('user1', 'incident-1', statement)
 })
 
 test('should call submitStatement', async () => {
   await service.submitStatement('user1', 'incident-1')
 
-  expect(incidentClient.submitStatement).toBeCalledTimes(1)
-  expect(incidentClient.submitStatement).toBeCalledWith('user1', 'incident-1')
+  expect(statementsClient.submitStatement).toBeCalledTimes(1)
+  expect(statementsClient.submitStatement).toBeCalledWith('user1', 'incident-1')
 })
 
 describe('validateSavedStatement', () => {
   test('valid statement', async () => {
-    incidentClient.getStatement.mockReturnValue({
+    statementsClient.getStatement.mockReturnValue({
       lastTrainingMonth: 1,
       lastTrainingYear: 2000,
       jobStartYear: 1998,
@@ -112,7 +110,7 @@ describe('validateSavedStatement', () => {
   })
 
   test('can cope with additional attributes', async () => {
-    incidentClient.getStatement.mockReturnValue({
+    statementsClient.getStatement.mockReturnValue({
       lastTrainingMonth: 1,
       lastTrainingYear: 2000,
       jobStartYear: 1998,
@@ -126,7 +124,7 @@ describe('validateSavedStatement', () => {
   })
 
   test('invalid statement', async () => {
-    incidentClient.getStatement.mockReturnValue({})
+    statementsClient.getStatement.mockReturnValue({})
 
     const errors = await service.validateSavedStatement('user-1', 1)
     expect(errors.map(error => error.href)).toEqual([
